Change: move naslvalue to nasl-syntax
With the movage we can use NaslValue in specialized crates and use them
in nasl-interpreter without introducing circular dependencies.
